Phase 1.5a (Task 39, direction Y): collapse identity user-ID int64→int32
The authforge::identity::* repository interfaces used int64_t for the internal
user ID, but the DB is int4 (users.id SERIAL, *_user_id INTEGER) and ORM
getters return int32_t -- so every Postgres implementation had to
static_cast<int32_t>(userId) before each ORM call (10+ sites). The legacy
oauth2::I* interfaces correctly used int32_t. Direction Y: align the new
interfaces with the DB/ORM truth (int32), not the reverse.

Changed (27 files, identity user-ID path only; timestamps left int64):
- 6 identity repo interfaces: IUser/IRole/ISubjectMapping/IMfa/IWebAuthn/
  ISocialAccount -- all userId params + UserData.id + OptionalIntCallback
- AuthService.h: AuthResult.internalId + getUserInfo
- Service layer: MfaService/WebAuthnService/SocialAuthService/SubjectResolver
- 3 Postgres impls: deleted static_cast intermediates (IdentityRepository 8
  methods, Mfa 6, WebAuthn 2); SocialAccount already int32 at ORM boundary
- 7 test fakes: InMemoryUserRepository/FakeMfaRepository/etc map keys + overrides

Verified: full build green; 76/76 identity unit tests pass (behavior
unchanged -- pure type alignment). githubId (GitHub external API id) correctly
kept int64.

+++ libs/identity/include/authforge/identity/AuthService.h
@@ -36,7 +36,7 @@ class IUserRepository;
  */
 struct AuthResult
 {
-    int64_t internalId = 0;      // Internal auto-increment ID (for DB operations)
+    int32_t internalId = 0;      // Internal auto-increment ID (for DB operations)
     std::string publicSub;       // Public UUID subject (for OAuth2 tokens, never expose internalId)
     bool emailVerified = false;  // Whether email is verified
     bool mfaEnabled = false;     // Whether MFA (TOTP) is enabled
@@ -101,7 +101,7 @@ class AuthService
      * @param callback Returns user info JSON (sub/name/email/roles) or
      * nullopt if not found.
      */
-    void getUserInfo(int64_t userId, std::function<void(std::optional<Json::Value>)> &&callback);
+    void getUserInfo(int32_t userId, std::function<void(std::optional<Json::Value>)> &&callback);
 
   private:
     std::shared_ptr<IUserRepository> userRepo_;
